Service / id - Integer로 수정

+++ backend/src/main/java/com/example/backend/member/dto/MemberListDto.java
@@ -18,7 +18,7 @@ public class MemberListDto {
 
     private String email;
 
-    public MemberListDto(Long id, String loginId, String name, String phone, String email) {
+    public MemberListDto(Integer id, String loginId, String name, String phone, String email) {
         this.id = id;
         this.loginId = loginId;
         this.name = name;
